+++ tests/hooks/useUpdateFileRenderElement.test.ts
@@ -16,31 +16,31 @@ describe('test useUpdateElement hook', () => {
     })
   })
 
-  it('should call update method with table props for File Render Element', () => {
+  it('should call update menthod with table props for File Render Element', () => {
     const props = { table: '', container: { type: Skyflow.ContainerType.REVEAL } }
     useUpdateFileRenderElement(props as any, elementMock as any)
     expect(updateMethodMock).toBeCalled()
   })
 
-  it('should call update method with column props File Render Element', () => {
+  it('should call update menthod with column props File Render Element', () => {
     const props = { column: '', container: { type: Skyflow.ContainerType.REVEAL } }
     useUpdateFileRenderElement(props as any, elementMock as any)
     expect(updateMethodMock).toBeCalled()
   })
 
-  it('should call update method with altText props File Render Element', () => {
+  it('should call update menthod with altText props File Render Element', () => {
     const props = { altText: '', container: { type: Skyflow.ContainerType.REVEAL } }
     useUpdateFileRenderElement(props as any, elementMock as any)
     expect(updateMethodMock).toBeCalled()
   })
 
-  it('should call update method with classes props File Render Element', () => {
+  it('should call update menthod with classes props File Render Element', () => {
     const props = { classes: '', container: { type: Skyflow.ContainerType.REVEAL } }
     useUpdateFileRenderElement(props as any, elementMock as any)
     expect(updateMethodMock).toBeCalled()
   })
 
-  it('should call update method with skyflowID props File Render Element', () => {
+  it('should call update menthod with skyflowID props File Render Element', () => {
     const props = { skyflowID: '', container: { type: Skyflow.ContainerType.REVEAL } }
     useUpdateFileRenderElement(props as any, elementMock as any)
     expect(updateMethodMock).toBeCalled()