+++ CHANGELOG.md
@@ -215,7 +215,7 @@ All notable changes to this project will be documented in this file.
 ## [1.4.0] - 2023-01-03
   
 ### Fixed
-- Fixed `Reveal` Element re-rendering
+- Fixed `Reveal` Element re-redering
 
 ## [1.3.1] - 2022-09-23
   
@@ -245,11 +245,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Minimum react version support for react SDK
+- Minimun react version support for react SDK
 ### Changed
 
 -  README.md file
 -  Sample App
 ## [1.0.0] - 2022-08-16
 
-- Initial Skyflow React SDK release
+- Inital Skyflow React SDK release